Making the deploy task work with any tree-ish
Instead of having a separate task for master and release, we can have a single task that takes a tree-ish argument.  As described in the comments, the checkout task does not do any merging (intentionally).  But all heads and tags are fetched first, so any commit hash or tag that has been pushed to the origin remote can be used.  To build the head from a remote branch, use the remote/branch syntax (e.g. origin/master).

The default task will deploy origin/master.  To deploy something else, provide an argument to the deploy task.  E.g.

    grunt deploy:r3.0.0-alpha.1